Add utils.js for utility methods for commands and ran npm init.
Created the utils.js file which will hold the main functions that will work with the CRUD operations between the command line and the file system. Also forgot to run npm init at the beginning of the project so went ahead and did that. Was able to get the create function working and saving notes to the file system.